avoid TTGO SD card pins

+++ components/nofrendo-esp32/gpiocontroller.c
@@ -40,14 +40,14 @@ void gpiocontrollerInit()
 	btn_config.mode = GPIO_MODE_INPUT;               //Set as Input
 	btn_config.pin_bit_mask = (uint64_t)             //Bitmask
 		(
-      ((uint64_t) 1 << UP)
-      | ((uint64_t) 1 << DOWN)
-      | ((uint64_t) 1 << LEFT)
-      | ((uint64_t) 1 << RIGHT)
-      | ((uint64_t) 1 << SELECT)
-      | ((uint64_t) 1 << START)
-      | ((uint64_t) 1 << A)
-      | ((uint64_t) 1 << B)
+      ((uint64_t)(((uint64_t)1) << UP))
+      | ((uint64_t)(((uint64_t)1) << DOWN))
+      | ((uint64_t)(((uint64_t)1) << LEFT))
+      | ((uint64_t)(((uint64_t)1) << RIGHT))
+      | ((uint64_t)(((uint64_t)1) << SELECT))
+      | ((uint64_t)(((uint64_t)1) << START))
+      | ((uint64_t)(((uint64_t)1) << A))
+      | ((uint64_t)(((uint64_t)1) << B))
     );
 	btn_config.pull_up_en = GPIO_PULLUP_ENABLE;      //Disable pullup
 	btn_config.pull_down_en = GPIO_PULLDOWN_DISABLE; //Enable pulldown
